Move OutlinedInfoIcon to folder 'Share'
